Medium: IPaddr2: do not send Gratuitous ARPs in the Cluster IP configuration
It is not necessary to send Gratuitous ARPs in the case of Cluster IP
because the MAC address is always the cluster MAC and does not change
during a failover, so nothing needs to be updated on the peers, and
it will not help to reduce the service down time.

Only when send_arp.libnet binary is used, they will be sent to retain
the old behavior.